Refactor FormInput to use .scss and BEM class naming convention
